refactor(acquisition): isolate OpenCLI support helpers
